Bug 3116 Added PURGE for listener on shutdown and checks if timers can be trigered.

+++ sxp-core/src/main/java/org/opendaylight/sxp/util/time/connection/DeleteHoldDownTimerTask.java
@@ -25,11 +25,13 @@ public class DeleteHoldDownTimerTask extends ConnectionTimerTask {
 
     @Override
     protected void performAction() {
-        LOG.info(owner + " Default{} [{}]", getClass().getSimpleName(), getPeriod());
-        try {
-            connection.purgeBindings();
-        } catch (Exception e) {
-            LOG.warn(owner + " {} {} | {}", getClass().getSimpleName(), e.getClass().getSimpleName(), e.getMessage());
+        if (connection.isStateDeleteHoldDown()) {
+            LOG.info(owner + " Default{} [{}]", getClass().getSimpleName(), getPeriod());
+            try {
+                connection.purgeBindings();
+            } catch (Exception e) {
+                LOG.warn(owner + " {} {} | {}", getClass().getSimpleName(), e.getClass().getSimpleName(), e.getMessage());
+            }
         }
         done();
     }
